backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery==5.5.1",
    "croniter==6.0.0",
    "django==5.2",
    "django-celery-beat>=2.8.1",
    "django-cors-headers==4.3.1",
    "django-environ==0.11.2",
    "django-filter==25.1",
    "djangorestframework==3.16.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "gevent==24.11.1",
    "google-genai>=1.32.0",
    "praw==7.8.1",
    "psycopg2-binary==2.9.10",
    "redis==5.2.1",
    "uvicorn>=0.35.0",
]
